Improve LogoutAsync method in CommandStorage
This commit enhances the LogoutAsync method in the CommandStorage service. It now checks whether a LoginId exists in the object storage before executing the delete operation, thus preventing possible errors due to non-existent LoginId.

+++ Spravy.Ui/Services/CommandStorage.cs
@@ -1112,7 +1112,11 @@ public static class CommandStorage
 
     private static async Task LogoutAsync(CancellationToken cancellationToken)
     {
-        await objectStorage.DeleteAsync(StorageIds.LoginId).ConfigureAwait(false);
+        if (await objectStorage.IsExistsAsync(StorageIds.LoginId).ConfigureAwait(false))
+        {
+            await objectStorage.DeleteAsync(StorageIds.LoginId).ConfigureAwait(false);
+        }
+
         await navigator.NavigateToAsync(ActionHelper<LoginViewModel>.Empty, cancellationToken).ConfigureAwait(false);
         await cancellationToken.InvokeUIAsync(() => mainSplitViewModel.IsPaneOpen = false);
     }
